converted datastructure from questions.json to category first
Wrote nunjucks code to convert the datastructure from questons.json to a datastructure where metacategories come first then categories then senses then questions.

Also fixed the indexes for the menu.

Also added a filter that can be used to add items to a dictionary. Turned out not to be needed, but potentially a useful filter.