File now saved as answers-spcr.json
application/json is specified as the mime type to save but whether you specify application/json or text/plain the end type is 'PlainTextType'.
When I tried to save as .sprc the saved type was document and then the load function would not load it in because it is looking for  'PlainTextType'.
If you change the extension later to .spcr it makes no difference - the file is still read in.