First version v1
First version to test out.

results.js is using textArea.innerText to store current progress and creates charts based on that.

There is a place for improvements. Checking if there was any change inside DOM could be enought, so Charts are not re-created from scratch every time a user changes view.